Plumb allow_cross_cell_resize into compute API resize()
This adds a stub method to the compute API which will
eventually perform a policy check and compute service
version check to determine if cross-cell resize is
allowed when performing a resize or cold migration. This
is always False for now but can be mocked for testing here
and in subsequent changes that add functional tests for
the cross-cell resize flow before the confirm/revert code
is ready.

For cold migration with a target host, validation logic
is split out into a separate method to add the cross-cell
validation of the target host since if cross-cell resize
is allowed, the target host could be in another cell from
where the instance currently exists.

Currently, the resize/migrate API actions are synchronous
until conductor RPC casts to prep_resize() on the selected
target compute service. This could be problematic during a
cross-cell resize since that flow is orchestrated from
(super)conductor and is synchronous, i.e. the compute services
do not RPC cast between each other like a traditional resize.
Until the 2.34 microversion, the live migrate API had the
same problem. If the request is allowed to perform a cross-cell
resize then we will RPC cast from API to conductor. The API
user still gets a 202 response either way.

Part of blueprint cross-cell-resize

Change-Id: I098f91d8c498e5a85266e193ad37c08aca4792b2