Trim fake_deserialize_context in test_conductor
The assertions on the user_id and project_id in the
fake_deserialize_context methods do not actually
cause tests to fail if they raise MismatchError
because oslo.messaging just swallows the exception
and logs the traceback. With enough of these getting
logged it will cause subunit parser failures in the
console output because the stream is too large.

This removes the fake method and just changes the
stub to be a lambda that returns self.context which
is what fake_deserialize_context was doing minus
the project_id/user_id assertions.

Change-Id: I26b201b410aa1d965dc7a6635c11c8b63b457a71
Partial-Bug: #1813147