Filter duplicates from compute API get_migrations_sorted()
During a cross-cell resize there are migration records with the
same UUID in multiple cells and we need to filter duplicates out
of the GET /os-migrations response. This adds logic to do that
which picks the record with the newest "updated_at" value.

As noted inline, an alternative to this would be leveraging the
Migration.hidden attribute which is not currently set to True
anywhere. If we used that attribute, we could handle duplicate
migrations the same way we handle duplicate instances, but we
would need to think about compatibility impacts to the
GET /os-migrations API since the "hidden" filter parameter on
that API is essentially useless today.

Part of blueprint cross-cell-resize

Change-Id: Ife9e5c77a02c8aac37311aac1e6d6814ff8de0e8