Use placement.inventory.inuse in report client
Since I9a833aa35d474caa35e640bbad6c436a3b16ac5e we've had the framework
for placement to return specific error codes allowing us to
differentiate among error conditions for oft-repeated status codes.
That change also included as its proof-of-concept a specific code for
the placement side of InventoryInUse - i.e. an attempt to delete an
inventory record for which there are existing allocations.

SchedulerReportClient was previously identifying this error condition by
parsing the text of the 409 response. With this change, it instead uses
the provided error code.

Change-Id: Ic621adcadf10cc607455eba48c4cb1882bde23fa