Start functional testing for cross-cell resize
This adds the beginning of functional tests for
cross-cell resize which get the instance to VERIFY_RESIZE
state and assert various aspects of the instance both in
the source cell and target cell.

The fake virt driver is updated to support creating a
server with device tags and also to have a host-specific
volume connector dict to aid in debugging test failures.

Confirm/revert resize flows will build on these tests.

Part of blueprint cross-cell-resize

Change-Id: I398aeb73da6bf09e177e5f1fdc668f59f06de8dd