Use the correct mdev allocated from the pGPU
In the next change in the series, we will use nested resource providers
for pGPUs. Accordingly, the allocation will be made on a child RP, not the
root RP.
This change aims to prepare this by allocating the right mediated device from the
same physical device.

Note that given we don't pass allocations when initializing the virt driver, in
case some mdevs are missing, we will recreate them blindly. Next change will
fix it.

Part of blueprint reshape-provider-tree
Part of blueprint vgpu-stein

Change-Id: Ia8f43199f70da9c5867acd9a76c43ca8c2c11c80