Make memtune parameters consistent with libvirt docs and code
LibvirtConfigGuestMemoryTune created a <memtune> element with
attribute "units", although the libvirt docs mention that it should be
"unit":

    https://libvirt.org/formatdomain.html#elementsMemoryTuning

It did not cause any issues because libvirt actually accepts both,
but domains defined using "units" are automatically converted to
use "unit".

Similarly, libvirt will automatically convert "K" to "KiB".

So rename the attribute from "units" to "unit", and use the same "KiB"
unit format as elsewhere in the code, e.g. LibvirtConfigCapsNUMACell
and LibvirtConfigGuestMemoryBackingPage.

Change-Id: I1e5a28cd30124a2b08ed732d45e5504b31c97284