RT: improve logging in _update_usage_from_migration
While working on some code which performs a resize_claim
I was finding that the actual new_flavor usage was not being
tracked on the destination host. It was a coding mistake
because I had not set instance.new_flavor before calling
resize_claim but that is extremely tightly coupled behavior
with how prep_resize works and the bowels of the
_update_usage_from_migration code works. This change simply
improves logging in the _update_usage_from_migration method
for debugging and updates the resize_claim docstring to
indicate the coupled nature of instance.new_flavor when calling
that method.

Change-Id: Ieb539c9a0cfbac743c579a1633234537a8e3e3ee