Make move_allocations handle empty source allocations
During a resize/cold migration, conductor moves the
instance allocations on the source compute node to the
migration record and then the scheduler creates allocations
for the instance on the destination compute node.

When resize fails, the allocations are swapped back from
the migration record to the instance so that the instance
consumer holds the same source node allocations as before
the resize, and the allocations against the destination
compute node are dropped.

Depending on the order of calls to do this revert/move,
the allocations may have already been moved so the
source node consumer (the migration record in this case)
will not have any allocations and if the target consumer
(the instance in this case) has allocations, we should
assume there is nothing to do and just return True.

Change-Id: I483750462b80aac609f0c2deb783b389ff5156c3