Stub out port binding create/delete in NeutronFixture
This is needed for functional tests that are going to
use the NeutronFixture for operations that require
dealing with port bindings.

The NeutronFixture.list_extensions stub is also fleshed
out to include a response with the binding-extended
extension which is what's used from the neutron.api.API
supports_port_binding_extension method and without that
we don't create/delete port bindings via that API anyway.

Change-Id: Ie3c19e524679cd8bb2314aeadfad9623d1fefc8e