api-ref: mark os-cells as deprecated
As part of an effort to scrub the nova docs of cells v1
references [1] this moves the os-cells API reference to
be grouped with the other deprecated APIs.

A section title is added, similar to the "Obsolete APIs"
section, for organizational purposes.

It should be noted that the os-cells APIs have not been
formally deprecated by capping them at some maximum
microversion like the other deprecated proxy APIs or
for the nova-network APIs, but cells v1 was deprecated
in Pike [2] and since the os-cells APIs are admin-only,
they can be considered effectively deprecated with the
cells v1 deprecation itself, and when we delete the
cells v1 service code the APIs will also be gone like
the other Obsolete APIs.

[1] http://lists.openstack.org/pipermail/openstack-discuss/2019-February/002569.html
[2] I1a173f7ce0715e684850e030c358e8175fa8724c

Change-Id: I833d2e8d8857448c2922df488ae9d9b425050ddc