Add confirm_snapshot_based_resize_at_source compute method
This adds the confirm_snapshot_based_resize_at_source compute
service method which will be RPC called from (super)conductor
when confirming a cross-cell resize.

This method is pretty basic and is similar to confirm_resize()
but not exactly. It cleans the guest disks, cleans up port
bindings and volume attachments, and frees up resource usage
in the resource tracker and placement.

The changes made to the instance object in the traditional
confirm_resize flow are not made here since conductor is going
to just destroy the instance in the source cell database
once we have cleaned up the source host.

Another difference is when this is called, it is with the copy
of the instance from the source cell database which still has
the host/node set pointing to the source host/node which is
unlike how a traditional resize works where the instance host/node
are set to the target host prior to confirmation.

Also, the resize.confirm.start/end notifications do not happen
here because conductor will send those notifications since the
confirmation process spans both the source and target cell.

Part of blueprint cross-cell-resize

Change-Id: If6cdd627f66a66c412f3c1d4997baab3552a9a9c