Add ConfirmResizeTask
This adds the ConfirmResizeTask task which will be called
from the API to orchestrate confirming a cross-cell resize.

This task is responsible for cleaning up the source host
and cell including destroying the guest from the source
hypervisor and hard deleting the instance-related records from
the source cell database. It also updates the information
about the instance and migration in the target cell database.

The resize.confirm.(start|end) notifications, which are
traditionally sent from the confirm_resize() method on the
source host, are sent from this task since confirming a cross
cell resize spans multiple cells.

Part of blueprint cross-cell-resize

Change-Id: I5b9d41ef34385689d8da9b3962a1eac759eddf6a