Use migration_status during volume migrating and retyping
When swapping volumes Nova has to identify if the swap itself is related
to an underlying migration or retype of the volume by Cinder.  Nova
would previously use the status of the volume to determine if the volume
was retyping or migrating.

However in the migration case where a volume is moved directly between
hosts the volume is never given a status  of migrating by Cinder leading
to Nova never calling the os-migrate_volume_completion cinder API to
complete the migration.

This change switches Nova to use the migration_status of the volume to
ensure that this API is called for both retypes and migrations.

Depends-On: https://review.openstack.org/#/c/639331/
Change-Id: I1bdf3431bda2da98380e0dcaa9f952e6768ca3af
Closes-bug: #1803961
(cherry picked from commit 53c3cfa7a02d684ce27800e22e00a816af44c510)
(cherry picked from commit 91282f879cb453e4baa6c6decb4fff849c7a1b2a)