Fix irrelevant-files for legacy-grenade-dsvm-neutron-multinode-live-migration
The change to irrelevant-files in I902e459093af9b82f9033d58cffcb2a628f5ec39
made it such that the job will *only* run on changes to the
nova/tests/live_migration/ path which was not the intent. This fixes
the regex to not run the job on changes to nova/tests/ unless those
changes are in the nova/tests/live_migration/ directory.

This would probably be easier to reason about long-term if the
nova/tests/live_migration/ directory was moved under nova/gate/
and then we could just use the standard irrelevant-files list. That
refactoring is left for another day though.

Change-Id: Iff2fca18a779e70deecf8e57279e2daf9bc5529a
Closes-Bug: #1816152