Add RevertResizeTask
This adds the RevertResizeTask which will be used to orchestrate
the revert of a cross-cell resize. It is responsible for updating
the instance mapping and hidden fields to hide the target cell
instance and show the source cell instance, clean up the target
host / cell, destroy the target cell instance, and (re)spawn the
instance in the source cell.

Part of blueprint cross-cell-resize

Change-Id: I7069f8807f353fa63c96494395bbbf7cc44562fb