Revert cross-cell resize from the API
This adds the logic to the API revertResize
operation such that if the migration is a cross-cell
resize the API will RPC cast to conductor to
revert the resize rather than directly to the dest
compute service like a traditional resize. Conductor
will then orchestrate the revert process between the
source and target cell.

Now that the API has revertResize plumbed this change
builds on the cross-cell resize functional tests by
reverting the resize for the volume-backed server test
and adds a revert from stopped image-backed test.

Part of blueprint cross-cell-resize

Change-Id: Iad211a54719c5fde039c48072c58f3af96afc8b8