Confirm cross-cell resize while deleting a server
This adds the code to handle the TODO in the API code
when deleting a server in 'resized' status. Like the
traditional confirm resize code used here, this will
synchronously RPC call conductor to orchestrate the
confirm operation which will delete the instance-related
records from the source cell before the API deletes the
instance in the target cell.

Part of blueprint cross-cell-resize

Change-Id: I50b23c99a017bbfed98fd072112f0d06dd303829