Use instance mappings to count server group members
This adds a get_count_by_uuids_and_user() method to the InstanceMapping
object and uses it to count instance mappings for the purpose of
counting quota usage for server group members. By counting server group
members via instance mappings, the count is resilient to down cells in
a multi-cell environment.

Part of blueprint count-quota-usage-from-placement

Change-Id: I3ff39d5ed99a68ad8678e5ff62b343f3018b4768