Use Selection object to fill request group mapping
To fill the request group - resource provider mapping nova needs to look
at the allocation that is made in placement as well as the traits of
the used providers. So far these were explicit placement calls. However
it turns out that the Selection object already contains the allocation
data so one of the extra placement calls can be removed.

This patch replaces the GET /allocations/<uuid> placement call from
conductor build instance codepath with the usage of the
Selection.allocation_request json blob.

Closes-Bug: #1819430

Change-Id: Iecbee518444bd282ce5f6fd019db41a322f76a83