Override the 'get' method in DriverBlockDevice class
The following methods are overridden in DriverBlockDevice class.

* __getattr__
* __getitem__

The 'get' method is not overridden.
The value cannot be got by the 'get' method
though the value can be got by '__getattr__' (e.g. bdm.volumd_id)
or '__getitem__' (e.g. bdm['volume_id']) method.
So override the 'get' method to fix the issue.

Change-Id: Ic665fc1956831110937d98553e526cb909e49997
Closes-Bug: #1816938