Add unit tests for missing VirtualInterface in 2.70 os-interface
This adds some simple unit tests for showing and listing attached
port interfaces where the corresponding VirtualInterface record
does not exist - which is possible for very old port attachments
since we started creating the VirtualInterface record for neutron
ports in the Newton release.

This is a follow up to I09420ff7134874dfe4dc399931c7740e81ecc2d0.

Part of blueprint expose-virtual-device-tags-in-rest-api

Change-Id: If14c4d6e97870de875b5dcd6bf68bd1b56ffad74