Parse <emulator> elements from virConnectGetCapabilities()
Extend the associated LibvirtConfigCapsGuest class so that it can
parse and store <emulator> elements in a way which can be reused in
subsequent commits by callers of virConnectGetDomainCapabilities().

Ongoing / future work such as SEV[0] and the move of the default machine
type to q35[1] will need to invoke virConnectGetDomainCapabilities()
in order to check whether SEV or q35 respectively are supported.
However this API requires 5 parameters[2]:

   1) the path to the emulator binary (e.g. qemu-system-x86_64)
   2) the domain architecture
   3) the machine type
   4) the virtualization type
   5) flags (not used yet)

4) is determined by CONF.libvirt.virt_type.  The caller of
virConnectGetDomainCapabilities() can decide which combinations of 2)
and 3) type to pass[3], but still needs to know 1).  Once 2) and 3)
are known, this can be determined from the <emulator> elements
returned by libvirt's virConnectGetCapabilities() API[4].  nova
already calls this and parses the response into a LibvirtConfigCaps
object.  However currently the parser ignores the <emulator> elements.
This patch fixes that.

[0] https://specs.openstack.org/openstack/nova-specs/specs/stein/approved/amd-sev-libvirt-support.html
[1] https://blueprints.launchpad.net/nova/+spec/handle-default-machine-type-as-q35
[2] https://libvirt.org/html/libvirt-libvirt-domain.html#virConnectGetDomainCapabilities
[3] Calling virConnectGetDomainCapabilities for every machine type is
    most likely overkill:
    https://bugzilla.redhat.com/show_bug.cgi?id=1683471#c7
[4] https://libvirt.org/formatcaps.html#elementGuest

blueprint: amd-sev-libvirt-support
blueprint: handle-default-machine-type-as-q35
Change-Id: Ibdc88bc5e0a214bff36a8351a1b76980f7015d16